refactor(keystore): rename `e2ei_intermediate_cert` -> `x509_intermediate_cert`
