+++ keystore/src/entities/mls/mod.rs
@@ -1,5 +1,4 @@
 mod e2ei_crl;
-mod e2ei_intermediate_cert;
 mod mls_pending_message;
 mod persisted_mls_group;
 mod persisted_mls_pending_group;
@@ -10,10 +9,10 @@ mod stored_epoch_encryption_keypair;
 mod stored_hpke_private_key;
 mod stored_keypackage;
 mod stored_psk_bundle;
+mod x509_intermediate_cert;
 mod x509_trust_anchor;
 
 pub use e2ei_crl::E2eiCrl;
-pub use e2ei_intermediate_cert::X509IntermediateCert;
 pub use mls_pending_message::{ConversationId, MlsPendingMessage, MlsPendingMessagePrimaryKey};
 pub use persisted_mls_group::{ParentGroupId, PersistedMlsGroup};
 pub use persisted_mls_pending_group::PersistedMlsPendingGroup;
@@ -24,4 +23,5 @@ pub use stored_epoch_encryption_keypair::StoredEpochEncryptionKeypair;
 pub use stored_hpke_private_key::StoredHpkePrivateKey;
 pub use stored_keypackage::StoredKeyPackage;
 pub use stored_psk_bundle::StoredPskBundle;
+pub use x509_intermediate_cert::X509IntermediateCert;
 pub use x509_trust_anchor::X509TrustAnchor;
